+++ apps/client/app/[schoolId]/page.tsx
@@ -1,25 +1,24 @@
+import { Button } from "components";
 import { testPageStyles as s } from "./TestPage.styles";
 
 export default function TestPage() {
   return (
     <div className={s.wrapper}>
-      <h1 className="text-head1 text-strong">두록(DoLog) 테마 테스트</h1>
+      <h1 className="text-head1 text-strong mb-6">두록(DoLog) 테마 테스트</h1>
 
       <div className="flex flex-wrap gap-4">
-        {/* 학교 메인 버튼 */}
-        <button className={s.mainButton}>학교 메인 버튼</button>
+        <Button>학교 메인 버튼</Button>
 
-        {/* CVA 적용 버튼 - outline 모드 */}
-        <button className={s.customButton({ intent: "outline" })}>
-          기본 버튼 (custom1 & 2)
-        </button>
+        <Button variant="secondary" size="md">
+          보조 버튼
+        </Button>
 
-        {/* CVA 적용 버튼 - primary 모드 */}
-        <button className={s.customButton({ intent: "primary" })}>
-          강조 버튼
-        </button>
+        <Button variant="outline" size="md">
+          기본 버튼 (Outline)
+        </Button>
 
-        <div className={s.chip}>Chip 확인</div>
+        <Button size="sm">작은 버튼</Button>
+        <Button size="lg">큰 버튼</Button>
       </div>
     </div>
   );
