+++ packages/components/src/ui/Button/Button.tsx
@@ -1,0 +1,15 @@
+import React from "react";
+import { buttonVariants, type ButtonVariantsProps } from "./Button.styles";
+import { cn } from "../lib/utils";
+
+export interface ButtonProps
+  extends React.ButtonHTMLAttributes<HTMLButtonElement>, ButtonVariantsProps {}
+
+export const Button = ({ className, variant, size, ...props }: ButtonProps) => {
+  return (
+    <button
+      className={cn(buttonVariants({ variant, size }), className)}
+      {...props}
+    />
+  );
+};
